+++ apps/web/src/app/[locale]/challenge/[slug]/getChallengeRouteData.ts
@@ -1,12 +1,15 @@
 import { type Session } from '@repo/auth/server';
 import { prisma } from '@repo/db';
 import { cache } from 'react';
+import { getAllFlags } from '~/utils/feature-flags';
 
 export type ChallengeRouteData = NonNullable<Awaited<ReturnType<typeof getChallengeRouteData>>>;
 
 // this is to data to populate the description tab (default tab on challenge page)
-export const getChallengeRouteData = cache((slug: string, session: Session | null) => {
-  return prisma.challenge.findFirstOrThrow({
+export const getChallengeRouteData = cache(async (slug: string, session: Session | null) => {
+  const featureFlags = await getAllFlags();
+
+  const challenge = await prisma.challenge.findFirstOrThrow({
     where: {
       slug,
       status: 'ACTIVE',
@@ -34,4 +37,29 @@ export const getChallengeRouteData = cache((slug: string, session: Session | nul
       },
     },
   });
+
+  /**
+   * Select the first track that the user is enrolled in for this challenge.
+   */
+  const track =
+    featureFlags?.enableInChallengeTrack && session
+      ? await prisma.track.findFirst({
+          where: {
+            trackChallenges: {
+              some: {
+                challengeId: challenge.id,
+                track: {
+                  enrolledUsers: {
+                    some: {
+                      id: session.user.id,
+                    },
+                  },
+                },
+              },
+            },
+          },
+        })
+      : null;
+
+  return { challenge, track };
 });
